Work around broken sz
Some broken protocol drivers (like sz) don't send a YModem terminate
at the end of a batch (or maybe just after a single file?) so SyncTERM
ends up trying to fall back from G to CRC-16 to 8-bit before ending.

This detects when the last known file was received, and if the YModem
header that follows is not correct, aborts the transfer.