Add a "fake mode" for output drivers that can't change screen size
Previously, the height was left the same, and the width was set to
the next-lowest "standard" size.  Now the cterm window is set to
the size of the requested mode and both top/bottom and side bars
are added to make the virtual screen the same as the requested one.

Using the "current" mode still retains the old behaviour.

Closes SF request 11