Fix menu draw when returning to dialing list from comment
Also note the spectacularily poorly named WIN_REDRAW and WIN_NODRAW
flags.  WIN_REDRAW indicates that this *is* a redraw of the menu,
which indicates that you don't need to redraw the menu (because it's
already correct on screen).  Weird, but not insane... WIN_NODRAW on
the other hand indicates this is *not* a redraw, and that the window
is incorrect on screen and you therefore *must* redraw the menu.

That is to say:
WIN_REDRAW prevents the window from being redrawn
WIN_NODRAW forces the window to be redrawn