Allow the client-name, server-name, and term-type to be passed as arguments
Optional, for easier use with game servers that take one of the 3 rlogin
negotation parameters as the name/code of the door to execute. A telgate
mode flag value argument must be provided (use 0 for default behavior)
if you want to provide any of the other arguments to override the defaults
(the user's alias, real name, and current detected terminal type).