SlyEdit v1.75: Refactored the quote selection menu and cross-post selection menus to use DDLightbarMenu instead of SlyEdit's own lightbar code.  This allowed the elimination of most of the rest of SlyEdit's own lightbar code, reducing the size of SlyEdit.js by about 40Kb.  It also provides some consistency in behavior by using a common menu component.
Also, to support some additional required behaviors by SlyEdit, added the following
'event' functions to the DDLightbarMenu class:
ValidateSelectItem(pItemRetval): For validating that the user can select an item.
Takes the item's return value, and returns a bool to indicate whether the item can
be selected.

OnItemSelect(pItemRetval, pSelected): A function that is called when an item is being
selected (or de-selected, when multi-select is enabled).  The parameters are the item's
return value and a boolean to indicate whether the item was selected or de-selected.

Also, added an explicit DoKeyDown() function to DDLightbarMenu to support scrolling
the menu down when desired (special case for SlyEdit's quote selection window).