Add -file=name|pattern option to specify files to add/update
postfile.js still should be used for adding a single file to a filebase,
but when updating a single file, this seemed like a simple enhancement.
It also allows filtering of files to be added by wildcard (e.g. *.zip),
so potentially useful for adding files too.

Address enhancement request #348